app/src/main/java/com/oleg/hubal/bankconverter/global/CurrencyJsonUtils.java
package com.oleg.hubal.bankconverter.global;

import java.io.IOException;

import okhttp3.OkHttpClient;
import okhttp3.Request;
import okhttp3.Response;

/**
 * Created by User on 18.01.2017.
 */

public class CurrencyJsonUtils {

    public static Response getResponseWithRequest(String requestUrl) throws IOException {
        OkHttpClient client = new OkHttpClient();

        Request request = new Request.Builder()
                .url(requestUrl)
                .build();

        Response response = client.newCall(request).execute();
        return response;
    }

}
